+++ src/components/Form/FormLabelText.tsx
@@ -1,0 +1,17 @@
+import classNames from 'classnames';
+import { forwardRef, HTMLProps } from 'react';
+
+export interface FormLabelTextProps extends HTMLProps<HTMLSpanElement> {
+  isRequired?: boolean;
+}
+
+export const FormLabelText = forwardRef<HTMLSpanElement, FormLabelTextProps>(
+  ({ children, className, isRequired, ...props }, ref) => (
+    <span className={classNames('label-text', className)} ref={ref} {...props}>
+      {children}{' '}
+      {isRequired === true ? <span className="font-normal">*</span> : null}
+    </span>
+  ),
+);
+
+FormLabelText.displayName = 'FormLabelText';